+++ ToDo.xcodeproj/xcuserdata/aleksandrandreev.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -39,22 +39,6 @@
       <BreakpointProxy
          BreakpointExtensionID = "Xcode.Breakpoint.FileBreakpoint">
          <BreakpointContent
-            uuid = "4139B76B-2D43-4A0B-8F49-F58A473B5E6E"
-            shouldBeEnabled = "No"
-            ignoreCount = "0"
-            continueAfterRunningActions = "No"
-            filePath = "ToDo/Model/ToDoItem/TodoItem + JSON parsing.swift"
-            startingColumnNumber = "9223372036854775807"
-            endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "50"
-            endingLineNumber = "50"
-            landmarkName = "parse(json:)"
-            landmarkType = "7">
-         </BreakpointContent>
-      </BreakpointProxy>
-      <BreakpointProxy
-         BreakpointExtensionID = "Xcode.Breakpoint.FileBreakpoint">
-         <BreakpointContent
             uuid = "43FD6505-17CD-4658-BEAD-FB8802668439"
             shouldBeEnabled = "No"
             ignoreCount = "0"
