+++ ToDo.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -1,0 +1,33 @@
+{
+  "originHash" : "4c70d8de049006c4eb387db02086eb79de34c12fd28210628ece9b182749b4f9",
+  "pins" : [
+    {
+      "identity" : "cocoalumberjack?tab=readme-ov-file",
+      "kind" : "remoteSourceControl",
+      "location" : "https://github.com/CocoaLumberjack/CocoaLumberjack?tab=readme-ov-file",
+      "state" : {
+        "revision" : "4b8714a7fb84d42393314ce897127b3939885ec3",
+        "version" : "3.8.5"
+      }
+    },
+    {
+      "identity" : "filecachepackage",
+      "kind" : "remoteSourceControl",
+      "location" : "https://github.com/BelyahRU/FileCachePackage",
+      "state" : {
+        "branch" : "main",
+        "revision" : "a346808edbd3fba80d0a0b1f379c137fec616732"
+      }
+    },
+    {
+      "identity" : "swift-log",
+      "kind" : "remoteSourceControl",
+      "location" : "https://github.com/apple/swift-log",
+      "state" : {
+        "revision" : "9cb486020ebf03bfa5b5df985387a14a98744537",
+        "version" : "1.6.1"
+      }
+    }
+  ],
+  "version" : 3
+}
