+++ ToDo/Model/Classes/Observable.swift
@@ -1,0 +1,31 @@
+//
+//  Observable.swift
+//  ToDo
+//
+//  Created by Александр Андреев on 05.07.2024.
+//
+
+import Foundation
+
+class Observable<T> {
+    
+    var value: T? {
+        didSet {
+            DispatchQueue.main.async {
+                self.listener?(self.value)
+            }
+        }
+    }
+    
+    init( _ value: T?) {
+        self.value = value
+    }
+    
+    
+    private var listener: ((T?) -> Void)?
+    
+    func bind( _ listener: @escaping ((T?) -> Void)) {
+        listener(value)
+        self.listener = listener
+    }
+}
