+++ ToDo/Model/Extension/Binding + Ext.swift
@@ -1,0 +1,17 @@
+//
+//  Binding + Ext.swift
+//  ToDo
+//
+//  Created by Александр Андреев on 29.06.2024.
+//
+
+import Foundation
+import SwiftUI
+extension Binding {
+    init(_ base: Binding<Value>) {
+        self.init(
+            get: { base.wrappedValue },
+            set: { newValue in base.wrappedValue = newValue }
+        )
+    }
+}
