+++ ToDo/Model/Extension/Date + Ext.swift
@@ -1,0 +1,29 @@
+//
+//  Date + Ext.swift
+//  ToDo
+//
+//  Created by Александр Андреев on 28.06.2024.
+//
+
+import Foundation
+extension Date {
+    
+    func formattedDate() -> String {
+        let dateFormatter = DateFormatter()
+        dateFormatter.dateFormat = "d MMMM yyyy"
+        dateFormatter.locale = Locale(identifier: "ru_RU")
+        return dateFormatter.string(from: self)
+    }
+    
+    func nextDayFormatted() -> Date {
+        let nextDay = Calendar.current.date(byAdding: .day, value: 1, to: self) ?? self
+        return nextDay
+    }
+    
+    func getDayMonthFormatted() -> String {
+        let dateFormatter = DateFormatter()
+        dateFormatter.dateFormat = "d MMMM"
+        dateFormatter.locale = Locale(identifier: "ru_RU")
+        return dateFormatter.string(from: self)
+    }
+}
