+++ ToDo/Model/ToDoItem/TodoItem + JSON parsing.swift
@@ -6,7 +6,9 @@
 //
 
 import Foundation
-extension TodoItem {
+import FileCachePackage
+
+extension TodoItem: JSONParsable {
     
     enum ToDoDictionaryKeys: String, CaseIterable {
         case id
@@ -19,7 +21,7 @@ extension TodoItem {
         case category
     }
     
-    var json: Any {
+    var json: [String : Any] {
         var dictionary = [String: Any]()
         
         dictionary[ToDoDictionaryKeys.id.rawValue] = id
@@ -45,7 +47,7 @@ extension TodoItem {
         return dictionary
     }
     
-    static func parse(json: Any) -> TodoItem? {
+    static func parse(json: [String : Any]) -> TodoItem? {
         
         guard JSONSerialization.isValidJSONObject(json) else { return nil }
         
@@ -108,4 +110,6 @@ extension TodoItem {
 
         return item
     }
+    
+   
 }
