+++ ToDo/View/CalendarView/CalendarVC + CollectionView.swift
@@ -77,11 +77,10 @@ extension CalendarViewController: UICollectionViewDelegate, UICollectionViewData
     //Айтем выделен
     func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
         // Выделяем выбранную ячейку
-        collectionView.selectItem(at: indexPath, animated: true, scrollPosition: .centeredHorizontally)
         guard let cell = collectionView.cellForItem(at: indexPath) as? ToDoCollectionViewCell else { return }
-        cell.isSelected = true
         
         self.tableView.scrollToRow(at: IndexPath(item: 0, section: indexPath.row), at: .top, animated: false)
+        collectionView.selectItem(at: indexPath, animated: false, scrollPosition: .centeredHorizontally) 
         
     }
